Simple
- No more extra branch for deploy.
- No more Heroku support.
- Only Render, Railway and VPS.
---------

+++ bot/modules/bot_settings.py
@@ -14,15 +14,23 @@ from dotenv import load_dotenv
 from pyrogram.filters import command, create, regex
 from pyrogram.handlers import CallbackQueryHandler, MessageHandler
 
-from bot import DATABASE_URL, GLOBAL_EXTENSION_FILTER, IS_PREMIUM_USER, LOGGER, MAX_SPLIT_SIZE, SHORTENER_APIS, SHORTENERES, Interval, aria2, aria2_options, aria2c_global, bot, categories, config_dict, download_dict, extra_buttons, get_client, list_drives, qbit_options, status_reply_dict_lock, user_data
-from bot.helper.ext_utils.bot_utils import get_readable_file_size, new_thread, set_commands, setInterval, sync_to_async
+from bot import (DATABASE_URL, GLOBAL_EXTENSION_FILTER, IS_PREMIUM_USER,
+                 LOGGER, MAX_SPLIT_SIZE, SHORTENER_APIS, SHORTENERES, Interval,
+                 aria2, aria2_options, aria2c_global, bot, categories,
+                 config_dict, download_dict, extra_buttons, get_client,
+                 list_drives, qbit_options, status_reply_dict_lock, user_data)
+from bot.helper.ext_utils.bot_utils import (get_readable_file_size, new_thread,
+                                            set_commands, setInterval,
+                                            sync_to_async)
 from bot.helper.ext_utils.db_handler import DbManger
 from bot.helper.ext_utils.task_manager import start_from_queued
 from bot.helper.mirror_utils.rclone_utils.serve import rclone_serve_booter
 from bot.helper.telegram_helper.bot_commands import BotCommands
 from bot.helper.telegram_helper.button_build import ButtonMaker
 from bot.helper.telegram_helper.filters import CustomFilters
-from bot.helper.telegram_helper.message_utils import editMessage, sendFile, sendMessage, update_all_messages
+from bot.helper.telegram_helper.message_utils import (editMessage, sendFile,
+                                                      sendMessage,
+                                                      update_all_messages)
 from bot.modules.rss import addJob
 from bot.modules.torrent_search import initiate_search_tools
 
@@ -34,12 +42,12 @@ default_values = {'AUTO_DELETE_MESSAGE_DURATION': 30,
                   'LEECH_SPLIT_SIZE': MAX_SPLIT_SIZE,
                   'RSS_DELAY': 900,
                   'STATUS_UPDATE_INTERVAL': 10,
-                  'SEARCH_LIMIT': 0}
+                  'SEARCH_LIMIT': 0,
+                  'UPSTREAM_BRANCH': 'jmdkh'}
 
 
 async def load_config():
 
-    
     BOT_TOKEN = environ.get('BOT_TOKEN', '')
     if len(BOT_TOKEN) == 0:
         BOT_TOKEN = config_dict['BOT_TOKEN']
@@ -105,10 +113,11 @@ async def load_config():
                 x = x.lstrip('.')
             GLOBAL_EXTENSION_FILTER.append(x.strip().lower())
 
+
     MEGA_EMAIL = environ.get('MEGA_EMAIL', '')
     MEGA_PASSWORD = environ.get('MEGA_PASSWORD', '')
-    if len(MEGA_EMAIL_ID) == 0 or len(MEGA_PASSWORD) == 0:
-        MEGA_EMAIL_ID = ''
+    if len(MEGA_EMAIL) == 0 or len(MEGA_PASSWORD) == 0:
+        MEGA_EMAIL = ''
         MEGA_PASSWORD = ''
 
     UPTOBOX_TOKEN = environ.get('UPTOBOX_TOKEN', '')
@@ -270,6 +279,14 @@ async def load_config():
         BASE_URL = ''
     else:
         await create_subprocess_shell(f"gunicorn web.wserver:app --bind 0.0.0.0:{BASE_URL_PORT} --worker-class gevent")
+
+    UPSTREAM_REPO = environ.get('UPSTREAM_REPO', '')
+    if len(UPSTREAM_REPO) == 0:
+        UPSTREAM_REPO = 'https://github.com/5hojib/Luna'
+
+    UPSTREAM_BRANCH = environ.get('UPSTREAM_BRANCH', '')
+    if len(UPSTREAM_BRANCH) == 0:
+        UPSTREAM_BRANCH = 'jmdkh'
 
     LOG_CHAT = environ.get('LOG_CHAT', '')
     LOG_CHAT = '' if len(LOG_CHAT) == 0 else int(LOG_CHAT)
@@ -453,6 +470,8 @@ async def load_config():
         "TELEGRAM_API": TELEGRAM_API,
         "TELEGRAM_HASH": TELEGRAM_HASH,
         "TORRENT_TIMEOUT": TORRENT_TIMEOUT,
+        "UPSTREAM_REPO": UPSTREAM_REPO,
+        "UPSTREAM_BRANCH": UPSTREAM_BRANCH,
         "UPTOBOX_TOKEN": UPTOBOX_TOKEN,
         "USER_SESSION_STRING": USER_SESSION_STRING,
         "USE_SERVICE_ACCOUNTS": USE_SERVICE_ACCOUNTS,
@@ -820,7 +839,14 @@ async def update_private_file(client, message, pre_message):
         elif file_name == 'config.env':
             load_dotenv('config.env', override=True)
             await load_config()
-        await message.delete()
+        if '@github.com' in config_dict['UPSTREAM_REPO']:
+            buttons = ButtonMaker()
+            msg = 'Push to UPSTREAM_REPO ?'
+            buttons.ibutton('Yes!', f"botset push {file_name}")
+            buttons.ibutton('No', "botset close")
+            await sendMessage(message, msg, buttons.build_menu(2))
+        else:
+            await message.delete()
     if file_name == 'rclone.conf':
         await rclone_serve_booter()
     await update_buttons(pre_message)
@@ -988,8 +1014,8 @@ async def edit_bot_settings(client, query):
         await event_handler(client, query, pfunc, rfunc)
     elif data[1] == 'editvar' and STATE == 'view':
         value = config_dict[data[2]]
-        if value and data[2] in ['DATABASE_URL', 'TELEGRAM_API', 'TELEGRAM_HASH',
-                                 'USER_SESSION_STRING', 'MEGA_API_KEY', 'MEGA_PASSWORD',
+        if value and data[2] in ['DATABASE_URL', 'TELEGRAM_API', 'TELEGRAM_HASH', 'UPSTREAM_REPO',
+                                 'USER_SESSION_STRING', 'MEGA_PASSWORD',
                                  'UPTOBOX_TOKEN'] and not await CustomFilters.owner(client, query):
             value = 'Only owner can see this!'
         elif len(str(value)) > 200:
@@ -1056,6 +1082,15 @@ async def edit_bot_settings(client, query):
     elif data[1] == 'push':
         await query.answer()
         filename = data[2].rsplit('.zip', 1)[0]
+        if await aiopath.exists(filename):
+            await (await create_subprocess_shell(f"git add -f {filename} \
+                                                   && git commit -sm botsettings -q \
+                                                   && git push origin {config_dict['UPSTREAM_BRANCH']} -qf")).wait()
+        else:
+            await (await create_subprocess_shell(f"git rm -r --cached {filename} \
+                                                   && git commit -sm botsettings -q \
+                                                   && git push origin {config_dict['UPSTREAM_BRANCH']} -qf")).wait()
+        await message.reply_to_message.delete()
         await message.delete()
